fix(tolk): Preserve full-width scalar bindings through execution
The native argument ABI accepts int64, but execution contexts, schedule
binding extraction and JIT replay narrowed values through OCaml int. Carry
int64 consistently across that shared protocol and Rune callers. Preserve
exact scalar bounds and keep the existing exact symbolic evaluator, with
checked host-sized results only at geometry and shape boundaries.

Validation: Tolk, Rune and Kaun checks and the GPT2 example build pass.
Native CPU and Metal suites execute both signed int64 endpoints, including
retained Metal mixed-width replay. UOp evaluation, scalar metadata, engine
capture and frontend JIT endpoint tests pass. CUDA endpoint coverage is
updated but physical CUDA execution remains an open hardware requirement.

+++ packages/rune/lib/jit.ml
@@ -4100,7 +4100,7 @@ type 'q compiled = {
   cp_zero_copy : bool;
   cp_ctx : Nx_effect.context;
   cp_linear : U.t;
-  cp_vars : (string * int) list;
+  cp_vars : (string * int64) list;
   cp_input_uops : U.t array;
       (* default storage for the schedule's PARAM slots; replay copies this
          array and supplies its input, output and shared-arena owners *)
